Add files via upload
Program using C++ to solve the complex Gordion knot problem, where there was a knot so intricate that it was prophesied that anyone who could untie it would become king of Asia. In simple words Gordian Knot is an Ancient Greek legend associated with Alexander the Great in Gordium in Phrygia, regarding a complex knot that tied an oxcart. Reputedly, whoever could untie it would be destined to rule all of Asia. In 333 BC Alexander was challenged to untie the knot. Instead of untangling it laboriously as expected, he dramatically cut through it with his sword, thus exercising another form of mental genius. It is thus used as a metaphor for a seemingly intractable problem which is solved by exercising brute force.